+++ CHANGES.md
@@ -4,6 +4,7 @@ JAMES Examples Module Changes
 Version 1.1.1 (under development)
 ---------------------------------
 
+ - Updated TSP examples (4A, 4B): temperature range and scaling.
  - Use US locale for all input and output.
 
 Version 1.1 (11/07/2015)